Add files via upload
Credit to the developpers at Upna Lab. The code was found at this link: https://www.instructables.com/Acoustic-Levitator/. We used the code for the BigLev model and changed the OCR1A variable value from 199 to 191. This OCR1A variable corresponds to the relative frequency output of the levitator as 80MHz/OCR1A will be quite close to the actual frequency outputted by the transducers. As OCR1A is lower, the frequency increases while the frequency decreases as the OCR1A value is raised.