3D Models used
The circuit holder and circuit enclosure files are the parts I 3D printed to contain my acoustic levitator circuiting for better organization and presentation. The BigLevLip file is a lip I attached to the outer rim of the BigLev models we built to turn them into standing, adjustable levitators with three supports.