Add files via upload
Here is the report I put together for my project. This focuses on mostly what we did and the differences from what Asier Marzo and some other colleagues have done. The links at the end of the report can be used to see more of the original designs and models where we obtained our material and inspiration.